# web/src/openapi/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Backend API for Society Management System. Development base URL: http://localhost:8000/api. Production base URL: http://localhost:8080/api.",
    "title": "Society Management API",
    "termsOfService": "http://localhost:8000/terms",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0"
  },
  "host": "localhost:8000",
  "basePath": "/api",
  "paths": {
...
    "/v1/auth/change-password": {
      "post": {
        "security": [
          {
            "AccessToken": []
          }
        ],
        "description": "Changes the authenticated user's password after verifying the current password, then clears auth cookies.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Change password",
        "parameters": [
          {
            "description": "Change password payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.ChangePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "schema": {
              "$ref": "#/definitions/models.ChangePasswordAPIResponse"
            }
          },
          "400": {
            "description": "Password mismatch, password reuse, or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/forgot-password": {
      "post": {
        "description": "Always returns a generic success message. If the email exists, sends a password reset OTP.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Request password reset OTP",
        "parameters": [
          {
            "description": "Forgot password payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.ForgotPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset instructions response",
            "schema": {
              "$ref": "#/definitions/models.ForgotPasswordAPIResponse"
            }
          },
          "400": {
            "description": "Invalid request or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "description": "Authenticates a verified active user and sets access_token and refresh_token HTTP-only cookies.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "parameters": [
          {
            "description": "Login payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "$ref": "#/definitions/models.LoginAPIResponse"
            }
          },
          "400": {
            "description": "Invalid request or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "401": {
            "description": "Invalid credentials",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "403": {
            "description": "Email not verified or account disabled",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
...
    "/v1/auth/profile": {
      "get": {
        "security": [
          {
            "AccessToken": []
          }
        ],
        "description": "Returns the authenticated user's profile. Does not refresh or rotate tokens.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Get current user profile",
        "responses": {
          "200": {
            "description": "Profile fetched successfully",
            "schema": {
              "$ref": "#/definitions/models.GetProfileAPIResponse"
            }
          },
          "401": {
            "description": "Missing, invalid, or expired access token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "403": {
            "description": "Account disabled",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "security": [
          {
            "RefreshToken": []
          }
        ],
        "description": "Validates the refresh_token cookie and reissues only the access_token cookie.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "responses": {
          "200": {
            "description": "Access token refreshed successfully",
            "schema": {
              "$ref": "#/definitions/models.RefreshTokenAPIResponse"
            }
          },
          "401": {
            "description": "Missing, invalid, or expired refresh token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "403": {
            "description": "Account disabled",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/register": {
      "post": {
        "description": "Creates a user account and sends an email verification OTP.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "parameters": [
          {
            "description": "Registration payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Account created successfully",
            "schema": {
              "$ref": "#/definitions/models.RegisterAPIResponse"
            }
          },
          "400": {
            "description": "Invalid request or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "409": {
            "description": "Email or phone already exists",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/resend-otp": {
      "post": {
        "description": "Sends a fresh email verification OTP for an unverified user.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Resend email verification OTP",
        "parameters": [
          {
            "description": "Resend OTP payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.ResendOTPRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OTP sent successfully",
            "schema": {
              "$ref": "#/definitions/models.ResendOTPAPIResponse"
            }
          },
          "400": {
            "description": "Invalid request or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "409": {
            "description": "Email already verified",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/reset-password": {
      "post": {
        "description": "Resets a user's password using a valid password reset OTP and clears auth cookies on success.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Reset password",
        "parameters": [
          {
            "description": "Reset password payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.ResetPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "schema": {
              "$ref": "#/definitions/models.ResetPasswordAPIResponse"
            }
          },
          "400": {
            "description": "Invalid OTP, password reuse, or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "404": {
            "description": "User or active OTP not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "429": {
            "description": "Too many OTP attempts",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/resident/register": {
      "post": {
        "description": "Creates a resident user for the public flat claim flow, marks the email verified, and sets auth cookies immediately. This does not create a society membership or flat resident.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Register resident without OTP",
        "parameters": [
          {
            "description": "Resident registration payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.ResidentRegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Resident account created successfully",
            "schema": {
              "$ref": "#/definitions/models.LoginAPIResponse"
            }
          },
          "400": {
            "description": "Invalid request or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "409": {
            "description": "Email or phone already exists",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
    "/v1/auth/verify-otp": {
      "post": {
        "description": "Verifies the registration email OTP and marks the user's email as verified.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Verify email OTP",
        "parameters": [
          {
            "description": "Email verification payload",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.VerifyOTPRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "schema": {
              "$ref": "#/definitions/models.VerifyOTPAPIResponse"
            }
          },
          "400": {
            "description": "Invalid OTP or validation error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "404": {
            "description": "User or OTP not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "429": {
            "description": "Too many OTP attempts",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponseDoc"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "models.ChangePasswordAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.MessageData"
        },
        "message": {
          "type": "string",
          "example": "Password changed successfully"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "models.ChangePasswordRequest": {
      "type": "object",
      "required": [
        "confirm_password",
        "current_password",
        "new_password"
      ],
      "properties": {
        "confirm_password": {
          "type": "string",
          "maxLength": 72,
          "minLength": 8
        },
        "current_password": {
          "type": "string",
          "maxLength": 72,
          "minLength": 8
        },
        "new_password": {
          "type": "string",
          "maxLength": 72,
          "minLength": 8
        }
      }
    },
...
    "models.ErrorResponseDoc": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/models.ErrorData"
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
...
    "models.ForgotPasswordAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.DevOTPMessageData"
        },
        "message": {
          "type": "string",
          "example": "if this email exists, password reset instructions have been sent"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "models.ForgotPasswordRequest": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 255
        }
      }
    },
    "models.GetProfileAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.UserData"
        },
        "message": {
          "type": "string",
          "example": "Profile fetched successfully"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
...
    "models.LoginAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.UserData"
        },
        "message": {
          "type": "string",
          "example": "Login successful"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "models.LoginRequest": {
      "type": "object",
      "required": [
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 255
        },
        "password": {
          "type": "string",
          "maxLength": 72,
          "minLength": 8
        },
        "phone_number": {
          "type": "string",
          "maxLength": 20
        }
      }
    },
...
    "models.RefreshTokenAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.MessageData"
        },
        "message": {
          "type": "string",
          "example": "Access token refreshed successfully"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "models.RegisterAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.UserMessageData"
        },
        "message": {
          "type": "string",
          "example": "Account created successfully"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "models.RegisterRequest": {
      "type": "object",
      "required": [
        "email",
        "first_name",
        "password",
        "phone_number"
      ],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 255
        },
        "first_name": {
          "type": "string",
          "maxLength": 100,
          "minLength": 2
        },
        "last_name": {
          "type": "string",
          "maxLength": 100
        },
        "password": {
          "type": "string",
          "maxLength": 72,
          "minLength": 8
        },
        "phone_number": {
          "type": "string",
          "maxLength": 20
        }
      }
    },
...
    "models.ResendOTPAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.DevOTPMessageData"
        },
        "message": {
          "type": "string",
          "example": "verification OTP sent successfully"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "models.ResendOTPRequest": {
      "type": "object",
      "required": [
        "email"
      ],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 255
        }
      }
    },
    "models.ResetPasswordAPIResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/models.MessageData"
        },
        "message": {
          "type": "string",
          "example": "Password reset successfully"
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    },
...
  },
  "securityDefinitions": {
    "AccessToken": {
      "description": "Access token stored as an HTTP-only cookie.",
      "type": "apiKey",
      "name": "access_token",
      "in": "cookie"
    },
    "OnboardingToken": {
      "type": "apiKey",
      "name": "onboarding_token",
      "in": "cookie"
    },
    "RefreshToken": {
      "description": "Refresh token stored as an HTTP-only cookie scoped to /api/v1/auth/refresh.",
      "type": "apiKey",
      "name": "refresh_token",
      "in": "cookie"
    }
  }
}